+++ CHANGELOG.md
@@ -21,9 +21,17 @@ Given a version number MAJOR.MINOR.PATCH, increment the:
 
 
 ## [Unreleased]
+### Added
+- Recipes section with 2 recipes
+- Breadcrumbs to the blog
+
 ### Changed
 - Updated renovate base branch
 - SemverRanges config added
+- Tweaked some of teh about page text
+
+### Fixed
+- Blog rich snippet data
 
 
 ## [v1.1.0] - 2022-10-26